services/tileserver/pyproject.toml
[tool.poetry]
authors = ["Daven Quinn <dev@davenquinn.com>"]
description = "Macrostrat tile server"
name = "macrostrat.tileserver"
packages = [{ include = "macrostrat_tileserver" }]
version = "2.1.0"

[tool.poetry.dependencies]
fastapi = "^0.111.0"
"macrostrat.database" = "^3.0"
"macrostrat.utils" = "^1.2.0"
pytest = "^7.2.1"
python = "^3.8"
python-dotenv = "^1.0.0"
timvt = { path = "./deps/timvt", develop = true }
typer = "^0.12.3"
uvicorn = "0.16.0"
titiler = "^0.18.5"
httpx = "^0.27.2"

[tool.poetry.extras]
# Extra dependencies to compile raster tiles
raster_tiles = ["mapnik"]

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
python-dotenv = "^1.0.1"
black = "^24.8.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.poetry.scripts]
tileserver = "macrostrat_tileserver.cli:_cli"

[tool.pytest.ini_options]
norecursedirs = "deps/*"
addopts = [
    "--import-mode=importlib",
]


